+++ plutarch-testlib/Plutarch/Test/Laws.hs
@@ -17,7 +17,6 @@ module Plutarch.Test.Laws (
 
 import Control.Applicative ((<|>))
 import Data.Kind (Type)
-import Plutarch.Internal.Numeric (PPositive, Positive)
 import Plutarch.LedgerApi.V1 qualified as V1
 import Plutarch.Prelude
 import Plutarch.Test.QuickCheck (checkHaskellEquivalent, checkHaskellEquivalent2)